+++ native-schema-provider/src/main/java/io/confluent/kafka/schemaregistry/builtin/utils/internal/Accessor.java
@@ -18,7 +18,7 @@
  * limitations under the License.
  */
 
-package io.confluent.kafka.schemaregistry.builtin.util.internal;
+package io.confluent.kafka.schemaregistry.builtin.utils.internal;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonNode;
